requirements.txt
fastapi
mangum
supabase
python-dotenv
pydantic
